+++ frontend-next/components/common/Home/MainContent/Posts.tsx
@@ -1,21 +1,19 @@
 import Button from "@components/common/Button/Button";
 import {SharedCard, HomeCard} from "layout/screen/PostCard";
-import { useState } from "react";
 import useSWR from "swr";
 import { get, post } from "utils/request";
-import { bool } from "yup";
 
-interface AllPostType {
+interface PostsProps {
   page: number,
   onFull: () => void,
 }
 
-const AllPost = ({
+const Posts = ({
   page,
   onFull,
-}: AllPostType) => {
+}: PostsProps) => {
 
-  const {data: posts, error: postsErrors} = useSWR<any>(`/api/v2/posts?page=${page}`, get, {
+  const {data: posts, error: postsErrors} = useSWR<any>(page ? `/api/v2/posts?page=${page}` : null, get, {
     shouldRetryOnError: false,
     revalidateOnFocus: false,
     onErrorRetry: (error) => {
@@ -57,37 +55,6 @@ const AllPost = ({
     })
   }
   </> 
-}
-
-const Posts = () => {
-  const [page, setPage] = useState<number>(1);
-  const posts = [];
-  const [fullPost, setFullPost] = useState<boolean>(false);
-
-  for (let i = 1; i <= page; i++ ) {
-    posts.push(<AllPost page={i} key={i} onFull={() => setFullPost(true)}/>);
-  };
-
-  return (
-    <div className="space-y-2">
-      <div className="space-y-10">
-        {posts}
-      </div>
-      {
-        !fullPost &&
-        <div className="flex justify-center">
-          <Button
-            variant="link"
-            className="text-primary font-semibold"
-            onClick={() => setPage(page + 1)}
-            disabled={fullPost}
-            >
-            See more
-          </Button>
-        </div>
-      }
-    </div>
-  );
 };
 
 export default Posts;
